Fixed olivetti-scale-width for integer heights
The function `olivetti-scale-width` incorrectly assumes that the height is
always a fraction. However, it can in very normal circumstances be an integer
such as 120 (see section 39.12.1 from the manual
(https://www.gnu.org/software/emacs/manual/html_node/elisp/Face-Attributes.html)).
In such a case, the old function scales the width by a number that is far too
large, which results in margins of width 0.

To see how this fails, set the height to something like 120 by using, for
example, buffer-face-mode.

I fixed this by scaling by height/100 if the height is an integer. If it is a
float or something else entirely, the new function behaves as the old.